refactor: move core libraries that aren't used by node back into router
