+++ core/src/lib.rs
@@ -1,3 +1,2 @@
-pub mod traits;
-pub mod types;
 pub mod validator;
+pub mod wire;
