+++ core/src/wire/mod.rs
@@ -1,15 +1,5 @@
-use bn254::{PublicKey, Signature};
 use bytes::{Buf, BufMut};
 use commonware_codec::{EncodeSize, Error, Read, ReadExt, ReadRangeExt, Write};
-
-/// Generic verification data that can be used by different verification methods
-#[derive(Debug, Clone)]
-pub struct VerificationData {
-    pub signatures: Vec<Signature>,
-    pub public_keys: Vec<PublicKey>,
-    /// Additional context data that might be needed by specific verification methods
-    pub context: Option<Vec<u8>>,
-}
 
 /// Aggregation protocol message and payload types
 #[derive(Clone, Debug, PartialEq)]
@@ -123,20 +113,5 @@ pub mod aggregation {
                 Payload::Signature(signature) => signature.encode_size(),
             }
         }
-    }
-}
-
-impl VerificationData {
-    pub fn new(signatures: Vec<Signature>, public_keys: Vec<PublicKey>) -> Self {
-        Self {
-            signatures,
-            public_keys,
-            context: None,
-        }
-    }
-
-    pub fn with_context(mut self, context: Vec<u8>) -> Self {
-        self.context = Some(context);
-        self
     }
 }
