--- router/src/creator/core.rs
@@ -1,1 +1,0 @@
-pub use commonware_avs_core::traits::Creator;
