+++ router/src/creator/mod.rs
@@ -1,8 +1,9 @@
 // Core traits and types
-pub mod core;
 pub mod queue;
+pub mod traits;
 
 pub use queue::{CreatorConfig, SimpleTaskQueue, TaskQueue};
+pub use traits::Creator;
 
 // Test module
 #[cfg(test)]
