+++ router/src/creator/traits.rs
@@ -2,8 +2,6 @@ use anyhow::Result;
 use async_trait::async_trait;
 use commonware_codec::{EncodeSize, Read, Write};
 
-use crate::types::VerificationData;
-
 #[async_trait]
 pub trait Creator: Send + Sync {
     /// Associated type for task metadata. Each creator implementation defines its own data structure.
@@ -21,26 +19,4 @@ pub trait Creator: Send + Sync {
     /// # Returns
     /// * `Self::TaskData` - The task metadata in the creator's specific format
     fn get_task_metadata(&self) -> Self::TaskData;
-}
-
-#[async_trait]
-pub trait VerificationExecutor<T = ()>: Send + Sync
-where
-    T: Send + Sync,
-{
-    async fn execute_verification(
-        &mut self,
-        payload_hash: &[u8],
-        verification_data: VerificationData,
-        task_data: Option<&T>,
-    ) -> Result<ExecutionResult>;
-}
-
-#[derive(Debug, Clone)]
-pub struct ExecutionResult {
-    pub transaction_hash: String,
-    pub block_number: Option<u64>,
-    pub gas_used: Option<u64>,
-    pub status: Option<bool>,
-    pub contract_address: Option<String>,
 }
