+++ router/src/lib.rs
@@ -1,6 +1,6 @@
 pub use commonware_avs_bindings as bindings;
-pub use commonware_avs_core::types as wire;
 pub use commonware_avs_core::validator;
+pub use commonware_avs_core::wire;
 pub mod creator;
 pub mod executor;
 pub mod ingress;
