+++ usecase-counter/router/src/executor.rs
@@ -4,7 +4,7 @@ use async_trait::async_trait;
 use commonware_avs_bindings::WalletProvider;
 use commonware_avs_bindings::blssigcheckoperatorstateretriever::BLSSigCheckOperatorStateRetriever::getNonSignerStakesAndSignatureReturn;
 use commonware_avs_bindings::counter::{self, Counter};
-use commonware_avs_core::traits::ExecutionResult;
+use commonware_avs_router::executor::ExecutionResult;
 use commonware_avs_router::executor::bls::{
     BlsSignatureVerificationHandler, convert_non_signer_data,
 };
