+++ usecase-counter/validator.rs
@@ -4,14 +4,14 @@ use alloy_primitives::U256;
 use alloy_provider::ProviderBuilder;
 use anyhow::Result;
 use commonware_avs_bindings::{ReadOnlyProvider, counter::Counter};
-use commonware_avs_core::types as wire;
+use commonware_avs_core::wire;
 use commonware_codec::{DecodeExt, ReadExt};
 use commonware_cryptography::sha256::Digest;
 use commonware_cryptography::{Hasher, Sha256};
 use commonware_eigenlayer::config::AvsDeployment;
 use std::{env, io::Cursor};
 
-use commonware_avs_core::validator::interface::ValidatorTrait;
+use commonware_avs_core::validator::ValidatorTrait;
 
 pub struct CounterValidator {
     counter: Counter::CounterInstance<(), ReadOnlyProvider>,
